Refactor app.py to improve variable naming and adjust module imports
A series of changes have been made to the app.py file to increase code clarity and maintainability. First, several lower-case constant values, such as `device` and `model_type`, have been capitalized to improve readability and follow accepted Python conventions for constant naming. The imports have been adjusted to remove `ingest_texts` which is no longer used, and replace it with the `LangchainRetriever` class from the `retrieve` module. The `make_qa` function was also replaced with `make_llm_chain` function for better QR performance. The `save` method is now used to store the split document texts. Also, a comment is added to remind adding file cache for the gradio version.